### 2024-03-14 #6482 @Kevin-Mattheus-Moerman
@jonbmartin thanks for this submission. I am a little worried about wether this work is in scope for JOSS in terms of our ["substantial scholarly effort criteria"](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort), in particular in relation to:

> As a rule of thumb, JOSS’ minimum allowable contribution should represent not less than three months of work for an individual.

This work appears to have been created by a single contributor/author, within about ~1 month. There are no forks/issues/PR's from others, and only you are listed under stars/watching. I'll trigger a scope check now by the editorial board, so they can help have a look. However, it would be great if you could comment on the above, thanks.
