### 2024-03-15 #6482 @jonbmartin
Hi @Kevin-Mattheus-Moerman, thank you for taking a look at our work. The vast majority of the code base was originally developed by the listed authors in two separate github projects, [SigPy](https://github.com/mikgroup/sigpy) and [SigPy.RF](https://github.com/jonbmartin/sigpy-rf). The code contributions are primarily documented here: https://github.com/jonbmartin/sigpy-rf/graphs/contributors.

Development took place over several years, beginning in 2019. Upon conference with the authors in late 2023, we felt that the pulse design tools had matured enough and were distinct enough from the mission of SigPy to be a stand-alone Python package, so I migrated that code to the new PulPy project. However, the code base represents the work of a number of individuals over many years, and between the two previous projects has received over 300 stars! We have not previously published anything relating to the pulse design tools outside of a conference abstract, and so felt it was time to describe them in JOSS.

This is described in the manuscript, but I would appreciate it if you or the reviewers have any suggestions on how to make this more clear to the users.
