### 2024-03-15 289c750
Update paper.bib DOI's
Added DOI's to those papers missing them, and fixed invalid DOIs. Remaining references without DOI's are either A) conference proceedings for which none have been created, B) software toolboxes for which none have been created, or C) old papers using the SICI referencing convention which is not recognized by the DOI database
