+++ docs/paper/paper.bib
@@ -5,6 +5,7 @@ number = {1},
 pages = {121--127},
 title = {{Multiband excitation pulses for hyperpolarized 13C dynamic chemical shift imaging}},
 volume = {194},
+doi = {10.1016/j.jmr.2008.06.010},
 year = {2008}
 }
 @article{AboSeada2019,
@@ -14,12 +15,13 @@ number = {1},
 pages = {362--376},
 title = {{Multiband RF pulse design for realistic gradient performance}},
 volume = {81},
+doi = {10.1002/mrm.27411},
 year = {2019}
 }
 @article{Bosch2023,
 abstract = {Objective: With modern optimization methods, free optimization of parallel transmit pulses together with their gradient waveforms can be performed on-line within a short time. A toolbox which uses PyTorch's autodifferentiation for simultaneous optimization of RF and gradient waveforms is presented and its performance is evaluated. Methods: MR measurements were performed on a 9.4T MRI scanner using a 3D saturated single-shot turboFlash sequence for B1+ mapping. RF pulse simulation and optimization were done using a Python toolbox and a dedicated server. An RF- and Gradient pulse design toolbox was developed, including a cost function to balance different metrics and respect hardware and regulatory limits. Pulse performance was evaluated in GRE and MPRAGE imaging. Pulses for non-selective and for slab-selective excitation were designed. Results: Universal pulses for non-selective excitation reduced the flip angle error to an NRMSE of (12.3±1.7)% relative to the targeted flip angle in simulations, compared to (42.0±1.4)% in CP mode. The tailored pulses performed best, resulting in a narrow flip angle distribution with NRMSE of (8.2±1.0)%. The tailored pulses could be created in only 66 s, making it feasible to design them during an experiment. A 90° pulse was designed as preparation pulse for a satTFL sequence and achieved a NRMSE of 7.1%. We showed that both MPRAGE and GRE imaging benefited from the pTx pulses created with our toolbox. Conclusion: The pTx pulse design toolbox can freely optimize gradient and pTx RF waveforms in a short time. This allows for tailoring high-quality pulses in just over a minute.},
 author = {Bosch, Dario and Scheffler, Klaus},
-doi = {10.1007/S10334-023-01134-7/FIGURES/8},
+doi = {10.1007/s10334-023-01134-7},
 file = {:home/jonathan/.local/share/data/Mendeley Ltd./Mendeley Desktop/Downloaded/Bosch, Scheffler - 2023 - FastPtx a versatile toolbox for rapid, joint design of pTx RF and gradient pulses using Pytorch's autodiffer.pdf:pdf},
 issn = {13528661},
 journal = {Magnetic Resonance Materials in Physics, Biology and Medicine},
@@ -90,6 +92,7 @@ url = {http://web.eecs.umich.edu/~fessler/irt/irt}
    month = {12},
    title = {AniRes2D: Anisotropic Residual-enhanced Diffusion for 2D MR Super-Resolution},
    url = {https://arxiv.org/abs/2312.04385v1},
+   doi = {10.1117/12.3008456},
    year = {2023},
 }
 @article{Shin2021,
@@ -142,6 +145,7 @@ url = {http://web.eecs.umich.edu/~fessler/irt/irt}
    city = {Oxford},
    publisher = {Elsevier Academic Press},
    title = {NMR Imaging in Biomedicine},
+   doi = {10.1007/BF02797382},
    year = {1982},
 }
 
@@ -337,7 +341,7 @@ year = {2006}
 @article{Pruessmann,
[…]
