+++ docs/paper/paper.bib
@@ -341,7 +341,7 @@ year = {2006}
 @article{Pruessmann,
 abstract = {New theoretical and practical concepts are presented for considerably enhancing the performance of magnetic resonance imaging (MRI) by means of arrays of multiple receiver coils. Sensitivity encoding (SENSE) is based on the fact that receiver sensitivity generally has an encoding effect complementary to Fourier preparation by linear field gradients. Thus, by using multiple receiver coils in parallel scan time in Fourier imaging can be considerably reduced. The problem of image reconstruction from sensitivity encoded data is formulated in a general fashion and solved for arbitrary coil configurations and k-space sampling patterns. Special attention is given to the currently most practical case, namely, sampling a common Cartesian grid with reduced density. For this case the feasibility of the proposed methods was verified both in vitro and in vivo. Scan time was reduced to one-half using a two-coil array in brain imaging. With an array of five coils double-oblique heart images were obtained in one-third of conventional scan time. Magn Reson Med 42:952-962, 1999.},
 author = {Pruessmann, Klaas P and Weiger, Markus and Scheidegger, Markus B and Boesiger, Peter},
-doi = {10.1002/MRM.},
+doi = {https://doi.org/10.1002/(SICI)1522-2594(199911)42:5<952::AID-MRM16>3.0.CO;2-S},
 file = {:home/jonathan/.local/share/data/Mendeley Ltd./Mendeley Desktop/Downloaded/Pruessmann et al. - Unknown - SENSE Sensitivity Encoding for Fast MRI(2).pdf:pdf},
 keywords = {MRI,SENSE,fast imaging,receiver coil array,sensitivity encoding},
 title = {{SENSE: Sensitivity Encoding for Fast MRI}},
