### 2024-03-15 #6482 @jonbmartin
Added DOI's to those papers missing them, and fixed invalid DOIs. Remaining references without DOI's are either A) conference proceedings for which none have been created, B) software toolboxes for which none have been created. The long invalid DOI with https:// prefix does resolve correctly when entered into the DOI database at doi.org; this is for an old paper which uses SICI referencing convetion.
