### 2024-03-19 #6482 @Kevin-Mattheus-Moerman
@jonbmartin thanks for those comments. The board has decided this may be in scope for JOSS. I will search for a handling editor.
