### 2024-03-21 #6482 @bwheelz36
Hi @emdupre
Yes, this sounds really interesting :-)
