### 2024-03-22 #6482 @emdupre
👋 Hi @kerkelae, @imr-framework,

Would you be willing to review _PulPy: A Python Toolkit for MRI RF and Gradient Pulse Design_ for [JOSS (the Journal of Open Source Software)](https://joss.theoj.org/)?

You can learn more about reviewing for JOSS -- including our conflict of interest policy -- [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).

The review takes place on GitHub and focuses on [the software](https://github.com/jonbmartin/pulpy) and this short paper:
https://github.com/openjournals/joss-papers/blob/joss.06482/joss.06482/10.21105.joss.06482.pdf

If you're available to review this work within the next month—or six weeks, at most—please let me know and I'll add you as a reviewer.

Once I have sufficient reviewers I'll open a dedicated review issue.

Thank you for considering !
