### 2024-04-05 #6586 @editorialbot
[REVIEW]: PulPy: A Python Toolkit for MRI RF and Gradient Pulse Design
**Submitting author:** @jonbmartin (<a href="http://orcid.org/0000-0002-9384-8056">Jonathan Martin</a>)
**Repository:** https://github.com/jonbmartin/pulpy
**Branch with paper.md** (empty if default branch):
**Version:** v1.8.8
