### 2024-04-15 #6586 @emdupre
👋  Hi everyone, and happy Monday !

@bwheelz36 and @curtcorum, I noticed you had not yet created your reviewer checklists, so I wanted to make sure you weren't encountering any issues in getting started.

Of course, if you have any questions or concerns, please don't hesitate to let me know ! And thank you again for agreeing to review this submission for _PulPy_ 💐
