### 2024-04-16 jonbmartin/pulpy#1 @bwheelz36
Initial JOSS review
This is my initial review, as per https://github.com/openjournals/joss-reviews/issues/6586#issuecomment-2060034328.
Following this initial review, I will likely have follow up questions since the initial review requests clearer documentation and examples.

# PULPY review

## Dev ops comments

- pip install went smoothly, as did installing developer requirements
- Tests exist, but do not appear to ever be executed - this one is a deal breaker.
  - Tests are not run as part of CI
  - test command `coverage run -m unittest` returns `Ran 0 tests in 0 seconds`
- Packaging a bit antiquated.
  - you have setup.cfg, setup.py, pyproject.toml, requirements.txt - all these can be replaced with just a single pyproject.toml. This isn't a deal breaker for me though, just something you may wish to consider fixing.
  - Similarly, if you move the commands in `run_tests.sh` into .github/workflows for me it's a bit cleaner - but again, not a deal breaker.
- Is pulpy.egg-info here on purpose? I suspect it was uploaded by accident?

## Code comments

- The fact the code passes checking by black and ruff is very encouraging; I didn't feel the need to dig too deeply into the source knowing that
- the file `__version__` does not need it's own file I don't think? could it be included in __init__ instead?

## Documentation Comments

Documentation is where I feel this package is really lacking, and is the other area I insist should be improved prior to publication. I dabble with MRI (I can run a sequence, I've never built one before) so I was pretty excited to see this package. But I have no idea where to start? I think you should ask yourselves who are my users? What will they want to do? And having answered those questions, provide a few examples of how this package helps them achieve those tasks.

As an example: could this package could help me generate a simple spin echo sequence for a given list of system hardware? Could it export that sequence such that I can run it on a siemens scanner?

Once this documentation piece is addressed, I may have more comments on the code itself since it would give me the opportunity to use it.

## Other packages/ work

- The authors mention that this work has been integrated into [[pypulseseq](https://github.com/imr-framework/pypulseq/blob/dev/setup.py)](https://github.com/imr-framework/pypulseq/blob/dev/setup.py). However, pulply does not seem to be a depdency of pypulseseq - could the authors clarify what they mean?
- I also don't understand how this differs from pulseseq/ pypulseseq? do all these packages do the same thing?
