### 2024-04-17 jonbmartin/pulpy#2 @curtcorum
JOSS review (in-progress)
https://github.com/openjournals/joss-reviews/issues/6586

Misc:

It looks like the logo image [pulpy logo image](https://github.com/jonbmartin/pulpy/blob/docs/figures/pulpy_logo_v2.png) did not load in [README.rst](https://github.com/jonbmartin/pulpy/blob/master/README.rst)

Functionality:

Documentation:

Software paper:
