### 2024-04-17 daa0f46
Update README.rst
https://github.com/openjournals/joss-reviews/issues/6586
