### 2024-04-30 #6586 @emdupre
👋  Hi everyone ! I just wanted to follow-up on the status of this review, as I see that you have started your checklists,  @curtcorum and @bwheelz36, but I'm not clear if you are still working through them or waiting on author / editor feedback.

Please let me know if you're hitting any specific blockers, or if you have any scheduling concerns I should be aware of.

I also noticed that https://github.com/jonbmartin/pulpy/issues/1 is still open ; @jonbmartin, I just wanted to confirm that you are welcome to address issues as they are created, rather than waiting until the initial review is complete ! This can help to keep the review process moving forward.

If you have any questions, of course, please don't hesitate to let me know !
