### 2024-05-07 7ee7e53
Update sim.py
add abrm_hp version with inclusion of b1 variation across space
