### 2024-05-20 9f48ef2
add discovery of tests to shell script
