+++ run_tests.sh
@@ -3,5 +3,6 @@ rm -rf docs/generated/
 black .
 isort .
 ruff .
-coverage run -m unittest
+python -m unittest discover tests
+coverage run --source=./tests -m unittest discover -s tests/
 sphinx-build docs -W -b linkcheck -d _build/doctrees _build/html
