### 2024-05-20 b0447ea
simplify root flip problem to suppress error from lejas
