### 2024-05-20 eb2641b
update documentation figures
