### 2024-05-20 0c7d789
remove pulpy.egg-info/ from git repository
