+++ pulpy.egg-info/SOURCES.txt
@@ -1,30 +1,0 @@
-LICENSE
-README.rst
-pyproject.toml
-setup.cfg
-setup.py
-pulpy/__init__.py
-pulpy/io.py
-pulpy/linop.py
-pulpy/sim.py
-pulpy/verse.py
-pulpy/version.py
-pulpy.egg-info/PKG-INFO
-pulpy.egg-info/SOURCES.txt
-pulpy.egg-info/dependency_links.txt
-pulpy.egg-info/top_level.txt
-pulpy/grad/__init__.py
-pulpy/grad/optim.py
-pulpy/grad/waveform.py
-pulpy/rf/__init__.py
-pulpy/rf/adiabatic.py
-pulpy/rf/b1sel.py
-pulpy/rf/multiband.py
-pulpy/rf/optcont.py
-pulpy/rf/ptx.py
-pulpy/rf/shim.py
-pulpy/rf/slr.py
-pulpy/rf/util.py
-tests/test_linop.py
-tests/test_sim.py
-tests/test_version.py
