### 2024-05-20 82091d6
rename action to reflect that sphinx is also built
