+++ .github/workflows/run_tests.yml
@@ -5,7 +5,7 @@
 # separate terms of service, privacy policy, and support
 # documentation.
 
-name: Run Tests
+name: Run Tests and Build Sphinx
 
 on:
   release:
