### 2024-05-20 d7744c4
add new workflow to run tests and lint. Allows one to add additional python versions if desired
