+++ .github/workflows/run-unittest-on-pr-open.yaml
@@ -25,9 +25,9 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install -r ./request_handler/requirements.txt
+          pip install -r ../../requirements.txt
           pip install boto3
       - name: Test with unittest
         run: |
           cd request_handler
-          python -m unittest discover -s ./test  -p 'test_*.py'
+          python -m unittest discover -s ../../tests  -p 'test_*.py'
