+++ .github/workflows/run_tests.sh
@@ -1,6 +1,5 @@
 set -e
 rm -rf docs/generated/
-black .
 isort .
 ruff .
 coverage run --source=./tests -m unittest discover -s tests/
