+++ .github/workflows/isort.yml
@@ -1,0 +1,12 @@
+name: Run isort
+on:
+  - push
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v3
+      - uses: isort/isort-action@v1
+        with:
+            requirements-files: "requirements.txt requirements-test.txt"
