+++ scripts/sphinx-build.sh
@@ -1,6 +1,3 @@
 set -e
 rm -rf docs/generated/
-isort .
-ruff .
-coverage run --source=./tests -m unittest discover -s tests/
 sphinx-build docs -W -b linkcheck -d _build/doctrees _build/html
