### 2024-05-20 0f40232
correct error in sphinx script
