+++ scripts/sphinx-build.sh
@@ -1,3 +1,4 @@
+cd ..
 set -e
 rm -rf docs/generated/
 sphinx-build docs -W -b linkcheck -d _build/doctrees _build/html
