+++ README.rst
@@ -121,6 +121,7 @@ To begin using `pulpy`, import the package in your Python script. For demo purpo
 then write to .i file:
 
 .. code-block:: python
+	
 	pp.ge_rf_params(pulse, dt=4e-6)
 	pp.signa(pulse, 'slr_ex')
 
