+++ .github/workflows/run-sphinx-build.yml
@@ -8,7 +8,8 @@
 name: Build Sphinx
 
 on:
-  - push
+  pull_request:
+    branches: [dev, master]
 
 jobs:
   example-job:
@@ -22,4 +23,4 @@ jobs:
       - name: Make the script files executable
         run: chmod +x sphinx-build.sh
       - name: Run a script
-        run: ./sphinx-build.sh
+        run: bash sphinx-build.sh
