+++ .github/workflows/run-sphinx-build.yml
@@ -22,5 +22,6 @@ jobs:
         uses: actions/checkout@v4  
       - name: Make the script files executable
         run: chmod +x sphinx-build.sh
-      - name: Run a script
-        run: bash sphinx-build.sh
+      - name: Run the script
+        run: |
+          ./sphinx-build.sh
