+++ .github/workflows/run-sphinx-build.yml
@@ -22,6 +22,10 @@ jobs:
         uses: actions/checkout@v4  
       - name: Make the script files executable
         run: chmod +x sphinx-build.sh
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install -r ./requirements.txt
       - name: Run the script
         run: |
           ./sphinx-build.sh
