+++ .github/workflows/run-sphinx-build.yml
@@ -25,7 +25,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install -r ../requirements.txt
+          pip install sphinx
       - name: Run the script
         run: |
           ./sphinx-build.sh
