### 2024-05-21 743d2ef
revert previous and add print to sphinx script
