### 2024-05-21 7ab7653
revert previous and add print to sphinx script
