+++ scripts/sphinx-build.sh
@@ -1,4 +1,5 @@
 cd ..
+echo "$PWD"
 set -e
 rm -rf docs/generated/
 sphinx-build docs -W -b linkcheck -d _build/doctrees _build/html
