+++ scripts/sphinx-build.sh
@@ -1,4 +1,4 @@
-cd ../..
+cd ..
 echo "$PWD"
 set -e
 rm -rf docs/generated/
