+++ .github/workflows/run-sphinx-build.yml
@@ -5,27 +5,15 @@
 # separate terms of service, privacy policy, and support
 # documentation.
 
-name: Build Sphinx
-
-on:
-  pull_request:
-    branches: [dev, master]
+name: "Pull Request Docs Check"
+on: 
+- pull_request
 
 jobs:
-  example-job:
+  docs:
     runs-on: ubuntu-latest
-    defaults:
-      run:
-        working-directory: ./scripts
     steps:
-      - name: Check out the repository to the runner
-        uses: actions/checkout@v4  
-      - name: Make the script files executable
-        run: chmod +x sphinx-build.sh
-      - name: Install dependencies
-        run: |
-          python -m pip install --upgrade pip
-          pip install sphinx
-      - name: Run the script
-        run: |
-          ./sphinx-build.sh
+    - uses: actions/checkout@v1
+    - uses: ammaraskar/sphinx-action@master
+      with:
+        docs-folder: "docs/"
