+++ .github/workflows/run-sphinx-build.yml
@@ -13,6 +13,10 @@ jobs:
   docs:
     runs-on: ubuntu-latest
     steps:
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        pip install sphinx
     - uses: actions/checkout@v1
     - uses: ammaraskar/sphinx-action@master
       with:
