+++ .github/workflows/run-sphinx-build.yml
@@ -16,7 +16,7 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install sphinx
+        pip install sphinx sphinx_rtd_theme
     - uses: actions/checkout@v1
     - uses: ammaraskar/sphinx-action@master
       with:
