### 2024-05-21 f4bee65
go back to old sphinx script version, put sphinx script in top level of directory
