### 2024-05-21 2d3f343
go back to old sphinx script version, put sphinx script in top level of directory
