+++ .github/workflows/run-sphinx-build.yml
@@ -5,19 +5,24 @@
 # separate terms of service, privacy policy, and support
 # documentation.
 
-name: "Pull Request Docs Check"
-on: 
-- pull_request
+name: Build Sphinx
+
+on:
+  pull_request:
+    branches: [dev, master]
 
 jobs:
   docs:
     runs-on: ubuntu-latest
     steps:
-    - name: Install dependencies
-      run: |
-        python -m pip install --upgrade pip
-        pip install sphinx sphinx_rtd_theme
-    - uses: actions/checkout@v1
-    - uses: ammaraskar/sphinx-action@master
-      with:
-        docs-folder: "docs/"
+      - name: Check out the repository to the runner
+        uses: actions/checkout@v4  
+      - name: Make the script files executable
+        run: chmod +x sphinx-build.sh
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install sphinx sphinx_rtd_theme
+      - name: Run the script
+        run: |
+          ./sphinx-build.sh
