### 2024-05-21 666a6c5
modify conf to mock import pulpy
