### 2024-05-21 8d90c9c
add some troubleshooting to conf.py
