+++ scripts/sphinx-build.sh
@@ -1,3 +1,4 @@
+cd ..
 echo "$PWD"
 set -e
 rm -rf docs/generated/
