+++ .github/workflows/run-sphinx-build.yml
@@ -14,6 +14,9 @@ on:
 jobs:
   docs:
     runs-on: ubuntu-latest
+    defaults:
+      run:
+        working-directory: ./scripts
     steps:
       - name: Check out the repository to the runner
         uses: actions/checkout@v4  
