### 2024-05-21 079a24c
add a bit more info on other notebooks
