### 2024-05-21 f7b505f
add a bit more info on other notebooks
