### 2024-05-21 d6b2456
add markdown README (fix issues with figues not displaying on github)
