+++ README.md
@@ -1,0 +1,224 @@
+---
+title: 'PulPy: Pulses in Python'
+---
+
+![PulPy logo](../docs/figures/pulpy_logo_v2.png){.align-center
+width="250px"}
+
+[Source Code](https://github.com/jonbmartin/pulpy) \|
+[Documentation](https://pulpy.readthedocs.io) \| [Demo
+Code](https://github.com/jonbmartin/pulpy-tutorials)
+
+[![Documentation Status](https://readthedocs.org/projects/pulpy/badge/?version=latest)](https://pulpy.readthedocs.io/en/latest/?badge=latest)
+
+Description
+===========
+
+Welcome to PulPy! PulPy is a Python package for RF pulse and gradient
+pulse design for MRI. PulPy focuses on the design of individual pulses
+rather than full pulse sequences (for this, see other packages, e.g.
+[PyPulSeq](https://github.com/imr-framework/pypulseq)).
+
+PulPy is the successor package to
+[SigPy.RF](https://github.com/jonbmartin/sigpy-rf), a sub package for RF
+pulse design nested inside of the
+[SigPy](https://github.com/mikgroup/sigpy) package for signal processing
+and image reconstruction. PulPy is designed to replicate functionality
+of SigPy.RF, but is able to stand alone as an independent Python
+package. It also places greater emphasis on MRI gradient waveform
+design, which is necessary for many classes of RF pulses but also has
+broader applicability. PulPy current relies on SigPy as a dependency and
+takes advantage of its\' abstraction classes. This dependency will be
+reduced in future releases.
+
+Installation
+============
+
+PulPy requires Python version \>= 3.8. The core module depends on
+`numba`, `numpy`, `sigpy`, `PyWavelets`, `scipy`, and `tqdm`.
+
+Via `pip`
+---------
+
+PulPy can be installed through `pip`:
+
+    pip install pulpy
+
+Developer Installation
+----------------------
+
+If you want to contribute to the PulPy source code, we recommend you
+install it with `pip` in editable mode:
+
+    cd /path/to/pulpy
+    pip install -e .
+
+To run tests and contribute, we recommend installing the following
+packages:
+
+    pip install coverage ruff sphinx sphinx_rtd_theme isort
+
+and run the script `run_tests.sh`.
+
+If you make modifications and would like to update the version number
+across the release, run bumpversion (where argument size is one of
+\'major\', \'minor\', \'patch\'):
+
+    bumpversion --current-version 1.8.1 size
+
+Getting Started
+---------------
+
+To begin using [pulpy]{.title-ref}, import the package in your Python
+script. For demo purposes, we\'ll also import the SigPy package\'s
+plotting functions:
+
+``` {.python}
+import pulpy as pp       # import full package
+import sigpy.plot as pl      # import a plotting function
+```
+
+1) RF Pulse Design and Simulation
+---------------------------------
+
+[pulpy]{.title-ref} allows you to easily design and simulate RF pulses.
+Here\'s an example of an SLR pulse design with multibanding:
+
+1a) define RF pulse parameters
+
+``` {.python}
+tb = 8      # RF pulse time-bandwidth product
+N = 512     # number of timepoints to design
+d1 = 0.01   # magnetization passband ripple level
+d2 = 0.01   # magnetization stopband ripple level
+p_type = 'ex'   # RF pulse type - a 90 degree excitation pulse
[…]
