+++ README.md
@@ -1,18 +1,17 @@
----
-title: 'PulPy: Pulses in Python'
----
+PulPy: Pulses in Python
+===========
+![PulPy logo](../docs/figures/pulpy_logo_v2.png)
 
-![PulPy logo](/docs/figures/pulpy_logo_v2.png)
-[Source Code](https://github.com/jonbmartin/pulpy) \|
-[Documentation](https://pulpy.readthedocs.io) \| [Demo
+[Source Code](https://github.com/jonbmartin/pulpy) |
+[Documentation](https://pulpy.readthedocs.io) | [Demo
 Code](https://github.com/jonbmartin/pulpy-tutorials)
 
-[![Documentation Status](https://readthedocs.org/projects/pulpy/badge/?version=latest)](https://pulpy.readthedocs.io/en/latest/?badge=latest)
+[![Documentation
+Status](https://readthedocs.org/projects/pulpy/badge/?version=latest)](https://pulpy.readthedocs.io/en/latest/?badge=latest)
 
-Description
-===========
+# Description
 
-Welcome to PulPy! PulPy is a Python package for RF pulse and gradient
+Welcome to PulPy\! PulPy is a Python package for RF pulse and gradient
 pulse design for MRI. PulPy focuses on the design of individual pulses
 rather than full pulse sequences (for this, see other packages, e.g.
 [PyPulSeq](https://github.com/imr-framework/pypulseq)).
@@ -26,24 +25,21 @@ of SigPy.RF, but is able to stand alone as an independent Python
 package. It also places greater emphasis on MRI gradient waveform
 design, which is necessary for many classes of RF pulses but also has
 broader applicability. PulPy current relies on SigPy as a dependency and
-takes advantage of its\' abstraction classes. This dependency will be
+takes advantage of its' abstraction classes. This dependency will be
 reduced in future releases.
 
-Installation
-============
+# Installation
 
 PulPy requires Python version \>= 3.8. The core module depends on
 `numba`, `numpy`, `sigpy`, `PyWavelets`, `scipy`, and `tqdm`.
 
-Via `pip`
----------
+## Via `pip`
 
 PulPy can be installed through `pip`:
 
     pip install pulpy
 
-Developer Installation
-----------------------
+## Developer Installation
 
 If you want to contribute to the PulPy source code, we recommend you
 install it with `pip` in editable mode:
@@ -60,31 +56,30 @@ and run the script `run_tests.sh`.
 
 If you make modifications and would like to update the version number
 across the release, run bumpversion (where argument size is one of
-\'major\', \'minor\', \'patch\'):
+'major', 'minor', 'patch'):
 
     bumpversion --current-version 1.8.1 size
 
-Getting Started
----------------
+## Getting Started
 
-To begin using [pulpy]{.title-ref}, import the package in your Python
-script. For demo purposes, we\'ll also import the SigPy package\'s
-plotting functions:
+To begin using <span class="title-ref">pulpy</span>, import the package
+in your Python script. For demo purposes, we'll also import the SigPy
+package's plotting functions:
 
-``` {.python}
+``` python
 import pulpy as pp       # import full package
 import sigpy.plot as pl      # import a plotting function
 ```
 
-1) RF Pulse Design and Simulation
----------------------------------
+## 1\) RF Pulse Design and Simulation
 
-[pulpy]{.title-ref} allows you to easily design and simulate RF pulses.
-Here\'s an example of an SLR pulse design with multibanding:
+<span class="title-ref">pulpy</span> allows you to easily design and
[…]
