+++ README.md
@@ -58,7 +58,7 @@ If you make modifications and would like to update the version number
 across the release, run bumpversion (where argument size is one of
 'major', 'minor', 'patch'):
 
-    bumpversion --current-version 1.8.1 size
+    bumpversion --current-version 1.8.2 size
 
 ## Getting Started
 
