+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend ="setuptools.build_meta"
 
 [project]
 name = "pulpy"
-version = "1.8.1"
+version = "1.8.2"
 authors = [
   { name="Jonathan Martin", email="jonathan.bach.martin@vumc.org" },
 ]
