+++ docs/paper/paper.md
@@ -81,7 +81,7 @@ MRI pulse design tools. We call this new package PulPy, short for Pulses in Pyth
 
 # Target Audience
 
-The PulPy toolbox has been developed for use by MRI researchers who are interested in pulse sequence design, MRI physics, signal processing, and optimization. We believe that it will serve as an essential building block for more general image acquisition tools which require specialized RF pulses. The toolbox has already been incorporated into open-source sequence development software such as Pulseq [@Layton2017] and PyPulseq [@SravanRavi2019] to provide RF pulses critical to the performance of various pulse sequences. Finally, end-to-end optimization of MRI pulse sequences and reconstructions is 
+The PulPy toolbox has been developed for use by MRI researchers who are interested in pulse sequence design, MRI physics, signal processing, and optimization. We believe that it will serve as an essential building block for more general image acquisition tools which require specialized RF pulses. The previous iteration of this toolbox, SigPy.RF, has already been incorporated into open-source sequence development software such as Pulseq [@Layton2017] and PyPulseq [@SravanRavi2019] to provide RF pulses critical to the performance of various pulse sequences. We feel that PulPy, with its' more specific focus on pulse design, will be able to even more easily integrated into other MRI acquisition software toolboxes, and we encourage other MRI software developers to incorporate PulPy as a component of MRI acquisition software.  Finally, end-to-end optimization of MRI pulse sequences and reconstructions is 
 being increasingly explored [@Radhakrishna2023; @Wang2022]; with the RF pulse and gradient waveform design functions
 provided, the PulPy package could facilitate this research. 
 
