### 2024-05-21 #6586 @jonbmartin
Hello all,

Thank you again for the initial review- it was very helpful, and I've made some changes to address the points raised. These correspond to a new release, release 1.8.2

**Comments response:**

1) unittests are now running, and are set to automatically run in response to opened pull requests via a github workflow.

2) Documentation has been increased. I have added more documentation in the README to help users get started. More comprehensive tutorials are linked to in a [separate tutorial repository ](https://github.com/jonbmartin/pulpy-tutorials). I would appreciate any further feedback on this from the reviewers.

3) @curtcorum very helpfully addressed an issue with figures not displaying properly on the github page with a pull request. Sphinx and Github have annoying contradictions in how they handle images which seems to be an [unresolved issue](https://stackoverflow.com/questions/27603008/getting-image-appearing-properly-in-readme-file-using-include-with-sphinx). My plan is to migrate to a more modern documentation package than Sphinx soon, but in the meantime I've created a separate README.rst and README.md for Sphinx and github respectively.

4)  @bwheelz36 made several helpful structural/cleanliness suggestions. I deleted some unnecessary clutter (e.g. pulpy.egg-info) and migrated the functionality in the shell script to Github workflows (one shell script is still called, but this is done by a workflow - keeps things a little more collected. This is again needed because of Sphinx, more motivation to move away from it in the future).
I have not yet taken the step of consolidating config files to a single pyproject.toml, but this was listed as a non-dealbreaker.

5) @bwheelz36 wanted clarity about the scope of the toolbox, especially vs. PyPulseq. PyPulseq is a more high level tool, good for building entire pulse sequences but with little focus on design/optimization of individual pulses. For example, I would recommend PyPulseq to build a full spin-echo pulse sequence. In contrast, PulPy would be a good tool to, for example, design a very specific refocusing RF pulse. That specific, optimized component could then be used in a higher-level toolbox like PyPulseq.

6) @bwheelz36 pointed out that PulPy is not a dependency of Pulseq or PyPulseq, as was stated in the paper. This is correct and was stated in error - the predecessor of PulPy, SigPy, was incorporated as a dependency in those packages. The manuscript has been updated to clarify this with: _"The previous iteration of this toolbox, SigPy.RF, has already been incorporated into open-source sequence development software such as Pulseq [@Layton2017] and PyPulseq [@SravanRavi2019] to provide RF pulses critical to the performance of various pulse sequences. We feel that PulPy, with its' more specific focus on pulse design, will be able to even more easily integrated into other MRI acquisition software toolboxes, and we encourage other MRI software developers to incorporate PulPy as a component of MRI acquisition software."_

I hope that this addresses the issues raised - please let me know if there is anything you would like further work on or clarification of.
