### 2024-05-22 jonbmartin/pulpy#1 @emdupre
Hello !

For tracking, I wanted to ping in @jonbmartin 's comments on the main JOSS review thread, many of which are intended to address the points raised in this issue : https://github.com/openjournals/joss-reviews/issues/6586#issuecomment-2123406700

@bwheelz36, please feel free to follow-up here as to whether you now consider this issue resolved, or if there are still points which remain unaddressed.
