### 2024-05-29 jonbmartin/pulpy#11 @bwheelz36
PyPi package does not install dependencies
Continuing review for https://github.com/openjournals/joss-reviews/issues/6586

I've noticed that although the package dependencies are listed in `requirements.txt`, this is not linked to the packaing of the code. As such, doing `pip install pulpy` does not install the dependencies required to run the code.
