### 2024-05-29 jonbmartin/pulpy#11 @bwheelz36
this also applies if one does `pip install -e .` as suggested in the readme
