### 2024-05-29 jonbmartin/pulpy#13 @bwheelz36
tests
Continuing review for https://github.com/openjournals/joss-reviews/issues/6586

- the readme still refers to a `runtests.sh` which no longer exists
- I suggest you update the tests to run on both PRs (already happening) and when AND on changes to main branch. the latter shows people very clearly at a glance that you have CI and tests running when they look at your code. Example trigger below:

```yaml
on:
  push:
    branches: [ main ]
  pull_request:
    branches: [ main ]
```

- Running the tests generates a deprecation warning: `deprecationWarning: You are passing weight=[1, 10.000000000000002] as a positional argument. Please change your invocation to use keyword arguments. From SciPy 1.14, passing these as positional arguments will result in an error.`
- coverage could be better. you don't need to fix this for me to approve JOSS paper...but you should do it anyway! Can use the below command to get a coverage report on command line: (requires coverage and pytest)
```
coverage run --source=pulpy -m pytest -v tests && coverage report -m
```
