### 2024-05-29 jonbmartin/pulpy#11 @bwheelz36
(this comes back to the packaging being pretty confusing. tbh, i'd recommend starting from scratch here and just using something like poetry)
