### 2024-06-02 #6586 @curtcorum
@emdupre @jonbmartin @bwheelz36

Review so far...

Some hopefully useful background, observations, comments and suggestions:

I have used the reconstruction functionality of the sigpy package for some time now (2-3 years). I had also sampled some of the RF functionality of sigpy.RF and found it useful.

I very much would like to see the pulpy package published in joss, and as a related goal, have a preview of capabilities, a robust set of requirements, setup instructions, tests, and user and api documentation. Sigpy itself is a very good example which embodies the above. It may be overkill in order to minimally meet the joss requirements but is still a great aspirational target.

The main difficulty with the review from my perspective has been getting the environment set up. As a relatively experienced python user, this is not a good sign. @bwheelz36 Has made numerous excellent suggestions to remedy this, but in my opinion it is still not there.

If not already, pulpy installation incuding automated tests should be hand tested on at least one type of clean system (not your existing development system) and ideally multiple platforms (windows, mac, linux). The CI functions of github could eventually help with this as @bwheelz36 has noted, but are not strictly needed to meet the joss requirements (again from my perspective), and also they are not necessarily the best way for initial system compatibility and environment debugging.

If some more effort can be made in this direction, I would be happy to evaluate the changes for an ubuntu 20.04 or 22.04 system in an isolated pip or conda environment. It is the main sticking point for meeting the joss functionality requirements from my perspective.

Documentation has improved generally, but the local installation and testing are still not there and also not for the development mode installation and instructions.

The functionality of the pulpy package as advertised is excellent! I was able to get pulpy running on binder with the changes in https://github.com/curtcorum/pulpy-tutorials/tree/ccjoss.
I'm looking forward to using it going forward as well as seeing it continue to be adopted by the community.

![adiabatic_notebook_binder](https://github.com/openjournals/joss-reviews/assets/1796102/1ab10daf-3b9d-402f-bec8-4abe96e08ba4)
