### 2024-06-10 b0db8d9
move to poetry! And create a poetry.lock file. Likely will not be used often but will provide a reference.
