### 2024-06-10 c7a673c
change a few minor details in pyproject.toml
