+++ .github/workflows/draft-pdf.yml
@@ -13,7 +13,7 @@ jobs:
         with:
           journal: joss
           # This should be the path to the paper within your repo.
-          paper-path: paper.md
+          paper-path: docs/paper/paper.md
       - name: Upload
         uses: actions/upload-artifact@v4
         with:
