### 2024-08-05 d0a757e
Update README.rst
testing replace path to figure with html
