### 2024-08-05 c3f7cee
Update README.rst
update all relative paths to absolute html based
