+++ README.rst
@@ -86,7 +86,7 @@ To begin using `pulpy`, import the package in your Python script. For demo purpo
 	pulse = pp.rf.slr.dzrf(N, tb, p_type, f_type, d1, d2, True)
 	pl.LinePlot(pulse,mode='r')     # plot the real component of the RF pulse
 
-.. image:: ../docs/figures/slr_pulse.png
+.. image:: https://github.com/jonbmartin/pulpy/blob/master/docs/figures/slr_pulse.png
    :align: center
    :width: 300
 
@@ -100,7 +100,7 @@ To begin using `pulpy`, import the package in your Python script. For demo purpo
 	mb_pulse = pp.rf.multiband.mb_rf(pulse, n_bands, band_sep, phs_type)
 	pl.LinePlot(mb_pulse)
 
-.. image:: ../docs/figures/multiband_pulse.png
+.. image:: https://github.com/jonbmartin/pulpy/blob/master/docs/figures/multiband_pulse.png
    :align: center
    :width: 300
 
@@ -115,10 +115,10 @@ To begin using `pulpy`, import the package in your Python script. For demo purpo
 	pl.LinePlot(Mxy_single_band, title='single band excitation')
 	pl.LinePlot(Mxy_multi_band, title='multi-band excitation')
 
-.. image:: ../docs/figures/single_band_excitation.png
+.. image:: https://github.com/jonbmartin/pulpy/blob/master/docs/figures/single_band_excitation.png
    :align: center
    :width: 300
-.. image:: ../docs/figures/multiband_excitation.png
+.. image:: https://github.com/jonbmartin/pulpy/blob/master/docs/figures/multiband_excitation.png
    :align: center
    :width: 300
 
@@ -146,7 +146,7 @@ then write to .i file:
         
         pl.LinePlot(trap, title='trapezoidal gradient')
 
-.. image:: ../docs/figures/trap_grad.png
+.. image:: https://github.com/jonbmartin/pulpy/blob/master/docs/figures/trap_grad.png
    :align: center
    :width: 300
 
@@ -166,7 +166,7 @@ to more complex time-varying waveforms (e.g. spiral gradient waveform):
         
         pl.LinePlot(np.transpose(g),mode='r', title='spiral gradient (1 axis plotted)')
 
-.. image:: ../docs/figures/spiral_waveform.png
+.. image:: https://github.com/jonbmartin/pulpy/blob/master/docs/figures/spiral_waveform.png
    :align: center
    :width: 300
 
