### 2024-08-05 377dcef
Update pyproject.toml with correct readme (rst)
