+++ pyproject.toml
@@ -4,7 +4,7 @@ version = "1.8.2"
 description = "PulPy: Pulses in Python. A package for MRI RF and gradient pulse design."
 authors = ["jonbmartin <jon.bach.martin@gmail.com>"]
 license = "MIT"
-readme = "README.md"
+readme = "README.rst"
 homepage = "https://github.com/jonbmartin/pulpy"
 repository = "https://github.com/jonbmartin/pulpy"
 classifiers = [
