### 2024-08-05 8141dfc
Update run-unittest-on-pr-open.yaml
update to run on push to master
