+++ .github/workflows/run-unittest-on-pr-open.yaml
@@ -5,9 +5,9 @@ name: ci-python-unittest
 
 on:
   pull_request:
-    branches: [dev, master]
+    branches: [dev, main]
   push:
-    branches: [master]
+    branches: [main]
 
 jobs:
   build:
