### 2024-08-05 jonbmartin/pulpy#13 @jonbmartin
Resolved. Have created a github action that run tests on both pull requests and pushes to master branch. Outdated references to readme.sh have been removed.
