### 2024-08-05 a927ac3
Update test_version.py
reduced independent version text referernces
