+++ tests/test_version.py
@@ -1,5 +1,5 @@
-import unittest
 import importlib.metadata
+import unittest
 
 from pulpy import version
 
@@ -9,4 +9,4 @@ if __name__ == "__main__":
 
 class TestVersion(unittest.TestCase):
     def test_version(self):
-        assert version.__version__ == importlib.metadata.version("mypackage")
+        assert version.__version__ == importlib.metadata.version("pulpy")
