### 2024-08-05 d6456e6
Delete tests/test_version.py
