+++ tests/test_version.py
@@ -1,12 +1,0 @@
-import importlib.metadata
-import unittest
-
-from pulpy import version
-
-if __name__ == "__main__":
-    unittest.main()
-
-
-class TestVersion(unittest.TestCase):
-    def test_version(self):
-        assert version.__version__ == importlib.metadata.version("pulpy")
