+++ pulpy/version.py
@@ -1,2 +1,2 @@
->>> import pkg_resources
->>> __version__ = pkg_resources.get_distribution('pulpy').version
+import pkg_resources
+__version__ = pkg_resources.get_distribution('pulpy').version
