+++ pulpy/version.py
@@ -1,1 +1,1 @@
-__version__=1.8.5
+__version__ = "1.8.5"
