### 2024-08-05 2c00f23
Update pyproject.toml version
