### 2024-08-05 jonbmartin/pulpy#11 @jonbmartin
Resolved! Dependencies install correctly now with 'pip install pulpy'
