### 2024-08-05 5d87fb4
Update slr.py firls to use weight as a kwarg for scipy compatibility
