+++ pulpy/rf/slr.py
@@ -136,7 +136,7 @@ def dzls(n=64, tb=4, d1=0.01, d2=0.01):
     m = [1, 1, 0, 0]
     w = [1, d1 / d2]
 
-    h = signal.firls(n + 1, f, m, w)
+    h = signal.firls(n + 1, f, m, weight=w)
     # shift the filter half a sample to make it symmetric, like in MATLAB
     c = np.exp(
         1j
@@ -250,8 +250,8 @@ def dz_gslider_b(n=128, g=5, gind=1, tb=4, d1=0.01, d2=0.01, phi=np.pi, shift=32
             m_sub = [1, 1, 0, 0, 0, 0]
             w = [1, 1, d1 / d2]
 
-            b_notch = signal.firls(n + 1, f, m_notch, w)  # the notched filter
-            b_sub = signal.firls(n + 1, f, m_sub, w)  # the subslice filter
+            b_notch = signal.firls(n + 1, f, m_notch, weight=w)  # the notched filter
+            b_sub = signal.firls(n + 1, f, m_sub, weight=w)  # the subslice filter
             # add them with the subslice phase
             b = np.add(b_notch, np.multiply(np.exp(1j * phi), b_sub))
             # shift the filter half a sample to make it symmetric,
@@ -337,13 +337,13 @@ def dz_gslider_b(n=128, g=5, gind=1, tb=4, d1=0.01, d2=0.01, phi=np.pi, shift=32
             * np.concatenate([np.arange(0, n / 2 + 1, 1), np.arange(-n / 2, 0, 1)])
         )
 
-        b_notch = signal.firls(n + 1, f, m_notch, w)  # the notched filter
+        b_notch = signal.firls(n + 1, f, m_notch, weight=w)  # the notched filter
         b_notch = sp.ifft(np.multiply(sp.fft(b_notch, center=False), c), center=False)
         b_notch = np.real(b_notch[:n])
         # hilbert transform to suppress negative passband
         b_notch = signal.hilbert(b_notch)
 
-        b_sub = signal.firls(n + 1, f, m_sub, w)  # the sub-band filter
+        b_sub = signal.firls(n + 1, f, m_sub, weight=w)  # the sub-band filter
         b_sub = sp.ifft(np.multiply(sp.fft(b_sub, center=False), c), center=False)
 
         b_sub = np.real(b_sub[:n])
@@ -447,8 +447,8 @@ def dz_hadamard_b(n=128, g=5, gind=1, tb=4, d1=0.01, d2=0.01, shift=32):
             / (2 * (n + 1))
             * np.concatenate([np.arange(0, n / 2 + 1, 1), np.arange(-n / 2, 0, 1)])
         )
-        bp = signal.firls(n + 1, f, mp, w)  # the positive filter
-        bn = signal.firls(n + 1, f, mn, w)  # the negative filter
+        bp = signal.firls(n + 1, f, mp, weight=w)  # the positive filter
+        bn = signal.firls(n + 1, f, mn, weight=w)  # the negative filter
 
         # combine the filters and demodulate
         b = sp.ifft(np.multiply(sp.fft(bp - bn, center=False), c), center=False)
@@ -882,7 +882,7 @@ def dz_ramp_beta(n, T, ptype, pbc, pbw, bs_offset, tb, d1, d2, dt):
     w = np.array([d1 / d2, 1, d1 / d2])  # band error weights
 
     # design the filter, firls requires odd numtaps, so design 1 extra & trim
-    b = signal.firls(n + 1, f, m, w)
+    b = signal.firls(n + 1, f, m, weight=w)
     b = b[0:n]
 
     # hilbert transformation to suppress negative passband, and demod to DC
