### 2024-08-05 6de0242
Update conf.py version
