### 2024-08-05 b67c77a
Update README.rst figs to absolute path
