### 2024-08-05 2ef1569
Update core.rst summarised items
