### 2024-08-05 c4ba4f1
Update core.rst revert to old structuring - was a sphinx bug
