+++ docs/core.rst
@@ -13,7 +13,7 @@ RF Linear Operators
     :toctree: generated
     :nosignatures:
 
-    PtxSpatialExplicit
+    pulpy.linop.PtxSpatialExplicit
 
 RF Pulse Simulation
 --------------------------
@@ -24,10 +24,10 @@ RF Pulse Simulation
     :toctree: generated
     :nosignatures:
 
-    abrm
-    abrm_nd
-    abrm_hp
-    abrm_ptx
+    pulpy.sim.abrm
+    pulpy.sim.abrm_nd
+    pulpy.sim.abrm_hp
+    pulpy.sim.abrm_ptx
 
 I/O
 --------------------------
@@ -38,7 +38,7 @@ I/O
     :toctree: generated
     :nosignatures:
 
-    siemens_rf
-    signa
-    ge_rf_params
-    philips_rf_params
+    pulpy.io.siemens_rf
+    pulpy.io.signa
+    pulpy.io.ge_rf_params
+    pulpy.io.philips_rf_params
