### 2024-08-05 36d4c98
Update pyproject.toml with new sphinx version
