+++ .github/workflows/run-unittest-on-pr-open.yaml
@@ -27,7 +27,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install -r ./requirements.txt
+          pip install .
           pip install boto3
       - name: Lint with Ruff
         run: |
