+++ .github/workflows/run-sphinx-build.yml
@@ -27,7 +27,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install -r ../requirements.txt
+          pip install ../.
           pip install sphinx sphinx_rtd_theme
       - name: Run the script
         run: |
