### 2024-08-05 87e5328
Update pyproject.toml
