+++ pyproject.toml
@@ -16,7 +16,7 @@ classifiers = [
 python = "^3.9"
 numpy = "^1.24.4"
 numba = "^0.59.1"
-PyWavelets = "^1.6.0"
+PyWavelets = "^1.5.0"
 scipy = "^1.13.1"
 tqdm = "^4.66.4"
 sigpy = "^0.1.26"
