### 2024-08-05 37f05cf
Update pyproject.toml remove numba and tqdm requirements (not necessary)
