+++ pyproject.toml
@@ -15,9 +15,7 @@ classifiers = [
 [tool.poetry.dependencies]
 python = "^3.9"
 numpy = "^1.24.4"
-numba = "^0.59.1"
 scipy = "^1.13.1"
-tqdm = "^4.66.4"
 sigpy = "^0.1.26"
 
 [tool.poetry.group.dev.dependencies]
