### 2024-08-05 de19979
Update pyproject.toml restore numba dependency
