### 2024-08-05 5fd196d
Update slr.py remez to use weight as kwargs
