+++ .github/workflows/run-unittest-on-pr-open.yaml
@@ -32,7 +32,7 @@ jobs:
       - name: Lint with Ruff
         run: |
             pip install ruff
-            ruff --output-format=github .
+            ruff check --output-format=github .
         continue-on-error: true
       - name: Test with unittest
         run: |
