+++ .github/workflows/run-unittest-on-pr-open.yaml
@@ -11,7 +11,7 @@ on:
 
 jobs:
   build:
-    runs-on: ubuntu-latest
+    runs-on: [ubuntu-latest, windows-latest]
     strategy:
       matrix:
         python-version: ["3.9", "3.10", "3.11", "3.12"]
