+++ .github/workflows/run-unittest-on-pr-open.yaml
@@ -11,10 +11,11 @@ on:
 
 jobs:
   build:
-    runs-on: [ubuntu-latest, windows-latest]
     strategy:
       matrix:
+        os: [ubuntu-latest, windows-latest]
         python-version: ["3.9", "3.10", "3.11", "3.12"]
+    runs-on: ${{ matrix.os }}
     env:
       AWS_ACCESS_KEY_ID: ${{ secrets.AWS_ACCESS_KEY_ID }}
       AWS_SECRET_ACCESS_KEY: ${{ secrets.AWS_SECRET_ACCESS_KEY }}
