### 2024-08-06 #6586 @jonbmartin
Hello all,

Thank you for the patience. I believe that I have now addressed all of the reviewers' comments and code critiques. I appreciate that they used issues/pull requests as a mechanism to point out specific issues, and I have addressed each of those individually. I will summarize here:

1) I have fully transitioned to using pyproject.toml and poetry. Unnecessary files (e.g. requirements.txt) have been eliminated.
2) I have made major improvements to the github actions. These now run unittests on windows and ubuntu using all supported python versions (3.9-3.12), do little things like isort, build the JOSS pdf regularly, and upload the python package.
3) Issues with pulpy not installing dependencies have been resolved, and has been hand tested on at least ubuntu systems. There is also coverage of this by the unittests of 2).
4) I have transitioned the demo jupyter notebooks to binder - everything runs in the browser! I also added a bit more documentation of each tutorial.
5) I have started the process of staging recipes for [conda](https://github.com/conda-forge/staged-recipes/pull/26624). This wasn't mentioned as a must but will definitely improve access and I will keep pushing on this.

One lingering issue unrelated to any reviewer critiques is the sphinx documentation  action failing. This seems to be related to a recently opened issue on Sphinx's end ([https://github.com/sphinx-doc/sphinx/issues/12589](https://github.com/sphinx-doc/sphinx/issues/12589)).

Please let me know if there are any other questions or issues. I will address quickly!

Thank you again for the very helpful review, I feel this has improved the package substantially.

Jonathan
