### 2024-08-21 #6586 @emdupre
👋 Hi @bwheelz36 and @curtcorum !

🌻 I know that we're well in the summer holiday season, but I just wanted to ping on this in case you're in-office. Please let me know if you have any feedback on the recent updates from @jonbmartin, or an estimated timeline for reviewing these ! And of course, if you have any questions on process that I can answer.

Thank you again !
