### 2024-08-26 jonbmartin/pulpy#18 @curtcorum
Development install of pulpy seems to be broken now.
#https://github.com/openjournals/joss-reviews/issues/6586

```
(pulpy) curt@green:~/src/pulpy$ pip install -e .
ERROR: File "setup.py" not found. Directory cannot be installed in editable mode: /mnt/md0/Champaign Dropbox/Curt Corum/src_green/pulpy
(A "pyproject.toml" file was found, but editable mode currently requires a setup.py based build.)

```
