### 2024-09-11 jonbmartin/pulpy#18 @emdupre
@jonbmartin I just wanted to follow up and make sure you had seen this issue. This is a blocker for finishing the JOSS review, so please let us know when you've had a chance to review it !
