### 2024-09-20 2b8ba13
Update README.rst
fix #18 by addressing some backend dependency suggestions
