### 2024-09-20 jonbmartin/pulpy#18 @jonbmartin
Hi @curtcorum and @emdupre thank you for raising this issue. So far I haven't been able to replicate. I am able to complete an editable install on all of the supported python versions. This could be an outdated version of pip (suggested [this thread](https://stackoverflow.com/questions/69711606/how-to-install-a-package-using-pip-in-editable-mode-with-pyproject-toml))

An example post-editable-install environment package list with python 3.10 is:

Package    Version Editable project location
---------- ------- --------------------------
llvmlite   0.42.0
numba      0.59.1
numpy      1.26.4
pip        24.2
pulpy      1.8.6   /home/jonathanmartin/pulpy
PyWavelets 1.7.0
scipy      1.14.1
setuptools 75.1.0
sigpy      0.1.26
tqdm       4.66.5
wheel      0.44.0

The thread linked above also suggests using setuptools v62.0.0 or newer. Downgrading to v61.0.0 and attempting to reinstall did not reproduce the error you describe, but making sure that is current can't hurt. I have updated the readme to suggest these backend packages. Let me know if this helps.
