### 2024-10-03 #6586 @jonbmartin
Hello @bwheelz36, @curtcorum, and @emdupre,

I just wanted to check back in on the review. @curtcorum I looked at the issue raised in https://github.com/openjournals/joss-reviews/issues/6586 and had difficulty replicating - I was wondering if @bwheelz36 has the same issue.

Best,
Jon
