### 2024-10-03 jonbmartin/pulpy#18 @curtcorum
@jonbmartin

I was able to get the development install to work with the following:

```
(pulpy) curt@green:~/src/pulpy$ mamba install pip

Looking for: ['pip']

nvidia/linux-64                                               No change
nvidia/noarch                                                 No change
frankong/noarch                                               No change
frankong/linux-64                                             No change
pytorch/noarch                                                No change
pytorch/linux-64                                              No change
pytorch-nightly/noarch                               1.2kB @   1.3kB/s  0.5s
conda-forge/noarch                                  16.7MB @  16.3MB/s  1.1s
conda-forge/linux-64                                38.6MB @  27.3MB/s  1.5s
pytorch-nightly/linux-64                           555.5kB @ 320.4kB/s  1.6s
Transaction

  Prefix: /home/curt/miniconda3/envs/pulpy

  Updating specs:

   - pip

  Package               Version  Build               Channel           Size
─────────────────────────────────────────────────────────────────────────────
  Install:
─────────────────────────────────────────────────────────────────────────────

  + _libgcc_mutex           0.1  conda_forge         conda-forge     Cached
  + ld_impl_linux-64       2.43  h712a8e2_1          conda-forge     Cached
  + ca-certificates   2024.8.30  hbcca054_0          conda-forge     Cached
  + libgomp              14.1.0  h77fa898_1          conda-forge     Cached
  + _openmp_mutex           4.5  2_gnu               conda-forge       24kB
  + libgcc               14.1.0  h77fa898_1          conda-forge     Cached
  + libzlib               1.3.1  hb9d3cd8_2          conda-forge       61kB
  + libgcc-ng            14.1.0  h69a702a_1          conda-forge     Cached
  + openssl               3.3.2  hb9d3cd8_0          conda-forge     Cached
  + libexpat              2.6.3  h5888daf_0          conda-forge     Cached
  + libsqlite            3.46.1  hadc24fc_0          conda-forge      865kB
  + tk                   8.6.13  noxft_h4845f30_101  conda-forge     Cached
  + libxcrypt            4.4.36  hd590300_1          conda-forge     Cached
  + libffi                3.4.2  h7f98852_5          conda-forge     Cached
  + bzip2                 1.0.8  h4bc722e_7          conda-forge     Cached
  + ncurses                 6.5  he02047a_1          conda-forge     Cached
  + libuuid              2.38.1  h0b41bf4_0          conda-forge     Cached
  + libnsl                2.0.1  hd590300_0          conda-forge     Cached
  + xz                    5.2.6  h166bdaf_0          conda-forge     Cached
  + readline                8.2  h8228510_1          conda-forge     Cached
  + tzdata                2024a  h8827d51_1          conda-forge     Cached
  + python               3.12.6  hc5c86c4_2_cpython  conda-forge       32MB
  + pip                    24.2  pyh8b19718_1        conda-forge     Cached
  + setuptools           75.1.0  pyhd8ed1ab_0        conda-forge     Cached
  + wheel                0.44.0  pyhd8ed1ab_0        conda-forge     Cached

  Summary:

  Install: 25 packages

  Total download: 32MB

─────────────────────────────────────────────────────────────────────────────

Confirm changes: [Y/n]
_openmp_mutex                                       23.6kB @ 299.1kB/s  0.1s
libzlib                                             61.0kB @ 679.4kB/s  0.1s
libsqlite                                          865.2kB @   3.5MB/s  0.2s
python                                              31.5MB @  36.8MB/s  0.9s

Downloading and Extracting Packages:

Preparing transaction: done
Verifying transaction: done
Executing transaction: done
```

```
(pulpy) curt@green:~/src/pulpy$ pip install -e .
Obtaining file:///mnt/md0/Champaign%20Dropbox/Curt%20Corum/src_green/pulpy
  Installing build dependencies ... done
  Checking if build backend supports build_editable ... done
  Getting requirements to build editable ... done
  Preparing editable metadata (pyproject.toml) ... done
Collecting numba<0.60.0,>=0.59.1 (from pulpy==1.8.6)
  Downloading numba-0.59.1-cp312-cp312-manylinux2014_x86_64.manylinux_2_17_x86_64.whl.metadata (2.7 kB)
Collecting numpy<2.0.0,>=1.24.4 (from pulpy==1.8.6)
  Downloading numpy-1.26.4-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata (61 kB)
Collecting scipy<2.0.0,>=1.13.1 (from pulpy==1.8.6)
  Downloading scipy-1.14.1-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata (60 kB)
Collecting sigpy<0.2.0,>=0.1.26 (from pulpy==1.8.6)
  Downloading sigpy-0.1.26-py3-none-any.whl.metadata (5.1 kB)
Collecting llvmlite<0.43,>=0.42.0dev0 (from numba<0.60.0,>=0.59.1->pulpy==1.8.6)
  Downloading llvmlite-0.42.0-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata (4.8 kB)
Collecting pywavelets (from sigpy<0.2.0,>=0.1.26->pulpy==1.8.6)
  Downloading pywavelets-1.7.0-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata (9.0 kB)
Collecting tqdm (from sigpy<0.2.0,>=0.1.26->pulpy==1.8.6)
  Downloading tqdm-4.66.5-py3-none-any.whl.metadata (57 kB)
Downloading numba-0.59.1-cp312-cp312-manylinux2014_x86_64.manylinux_2_17_x86_64.whl (3.7 MB)
Downloading numpy-1.26.4-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl (18.0 MB)
Downloading scipy-1.14.1-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl (40.8 MB)
Downloading sigpy-0.1.26-py3-none-any.whl (108 kB)
Downloading llvmlite-0.42.0-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl (43.8 MB)
Downloading pywavelets-1.7.0-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl (4.5 MB)
Downloading tqdm-4.66.5-py3-none-any.whl (78 kB)
Building wheels for collected packages: pulpy
  Building editable for pulpy (pyproject.toml) ... done
  Created wheel for pulpy: filename=pulpy-1.8.6-py3-none-any.whl size=15791 sha256=d345139c72193ef5d6e6b3824f09a5e307829edcb5a9001fb7cf8c0e28b28455
  Stored in directory: /tmp/pip-ephem-wheel-cache-6p_o48hz/wheels/32/8d/c8/2a36dd1a404379bbd9f8f2494f090b7dedf6243d8e6e390442
Successfully built pulpy
Installing collected packages: tqdm, numpy, llvmlite, scipy, pywavelets, numba, sigpy, pulpy
Successfully installed llvmlite-0.42.0 numba-0.59.1 numpy-1.26.4 pulpy-1.8.6 pywavelets-1.7.0 scipy-1.14.1 sigpy-0.1.26 tqdm-4.66.5
```
I've checked off the remaining issues in the review.

Congrats and I consider the review complete!
