### 2024-10-05 jonbmartin/pulpy#20 @bwheelz36
poetry.lock out of date
cloning this project and doing `poetry install` fails because the poetry.lock is out of sync with `pyproject.toml`.
Easily fixed by doing `poetry lock` or deleting `poetry.lock`. but should be fixed.
I would consider carefully whether you actually want to upload `poetry.lock` or not
