### 2024-10-05 jonbmartin/pulpy#21 @bwheelz36
poetry installof development dependency fails
```
The current project's supported Python range (>=3.9,<4.0) is not compatible with some of the required packages Python requirement:
  - sphinx requires Python >=3.10, so it will not be satisfied for Python >=3.9,<3.10
```
