### 2024-10-05 #6586 @bwheelz36
I've added a few more minor issues:

- https://github.com/jonbmartin/pulpy/issues/21
- https://github.com/jonbmartin/pulpy/issues/20

A substantial amount of time has passed since I agreed to review this work.
My initial review focused on the code packaging, test execution, and documentation. I believe that all of these elements have seen substantial improvement. I'm happy with where the work is from this perspective.

I've ticked off all review checkboxes except:

- [ ] Reproducibility: If the paper contains original results, results are entirely reproducible by reviewers. If the paper contains no original results, please check this item.

could you post the code need to reproduce figure 1, and I will verify I can do so.
