### 2024-10-09 7deee3e
update dependencies python >3.10, remove sphinx-rtd-theme
