+++ pyproject.toml
@@ -13,7 +13,7 @@ classifiers = [
 ]
 
 [tool.poetry.dependencies]
-python = "^3.9"
+python = "^3.10"
 numpy = "^1.24.4"
 numba = "^0.59.1"
 scipy = "^1.13.1"
@@ -23,7 +23,6 @@ sigpy = "^0.1.26"
 coverage = "^7.5.3"
 ruff = "^0.4.8"
 Sphinx = "^8.0.2"
-sphinx-rtd-theme = "^2.0.0"
 isort = "^5.13.2"
 
 [build-system]
