### 2024-10-09 2d013cf
Update README.rst
reflecting dependency change
