### 2024-10-09 jonbmartin/pulpy#21 @jonbmartin
Moved python dependency forward to >=3.10, resolved
