### 2024-10-10 #6586 @emdupre
@jonbmartin, please note the remaining author tasks described in [this comment](https://github.com/openjournals/joss-reviews/issues/6586#issuecomment-2405515124).

If you need any guidance in addressing these items, of course, please just let me know !
