### 2024-10-21 b0c42d4
Update paper.md with new affiliations and orcid
