### 2024-10-21 5bb76c8
Update run-unittest-on-pr-open.yaml
remove outdated python version
