### 2024-10-21 bb89121
Update pyproject.toml
