### 2024-10-22 #6586 @emdupre
Thank you, @jonbmartin ! I do have a few editorial requests :

- The title of the archive is currently "PulPy python package" ; could you please amend it to match the paper title, "PulPy: A Python Toolkit for MRI RF and Gradient Pulse Design" ? You should be able to do this without creating a new DOI by [editing the Zenodo metadata.](https://help.zenodo.org/docs/deposit/manage-records/#edit). If you encounter any issues, please let me know !
- In the author author affiliations (both on Zenodo and in the paper), could you please expand all acronyms ? For example, "USA" as "United States of America"

Specifically on the paper :
- For all appearances of `e.g.`, please add a comma after its occurrence. That is, `e.g.,`
- On L41, it seems that the `Soher2023` reference is not expanding correctly.
- On L42, I would assume that the Artiges reference should be in parentheses, rather than only its date in parentheses, since the authors are not discussed in text.
- Could you please change the citations in Table 1 and on L72 to avoid the double parentheses ? You should be able to do this using `[but see @citation]` rather than `(but see @citation)`.

On Table 1:
- Could you please add a caption for Table 1 ?

On Figure 1:
- On L33, `Figure ??` is referenced. Could you please correct this to cross reference ?
