### 2024-10-22 #6586 @emdupre
@jonbmartin could you please check the `SKIP DOIs` to confirm that indeed these references do not have associated DOIs ?

Otherwise, several of the references appear malformed. For example, L107 shows an unpaired parentheses and L158 shows a lone $B_1^+$. Could you please confirm the reference formatting ?
