### 2024-10-22 b5d2bf4
Update paper.md with expanded affiliations
