### 2024-10-22 8b6848d
Update paper.md (e.g. grammatical)
