### 2024-10-22 3511212
Update paper.md fix artiges reference collection
