+++ docs/paper/paper.md
@@ -68,7 +68,7 @@ Preliminary development of this toolbox was presented in reference [@Martin2020a
 
 
 # Statement of need
-The field of magnetic resonance imaging is currently experiencing rapid growth in available open source imaging tools. Tools have been made freely available for MRI hardware development [@Amrein2022; @Anand2018], system simulation [@Villena2014; @Stocker2010], pulse sequence programming [@Layton2017], image reconstruction [@Ong2019; @Uecker2015], and post-processing and analysis [@Avants2014; @Duval2018; Soher2023]. The great increase in open-source tools has helped enable fully open-source imaging systems [@Arndt2017; @Artiges2024]. 
+The field of magnetic resonance imaging is currently experiencing rapid growth in available open source imaging tools. Tools have been made freely available for MRI hardware development [@Amrein2022; @Anand2018], system simulation [@Villena2014; @Stocker2010], pulse sequence programming [@Layton2017], image reconstruction [@Ong2019; @Uecker2015], and post-processing and analysis [@Avants2014; @Duval2018; @Soher2023]. The great increase in open-source tools has helped enable fully open-source imaging systems [@Arndt2017; @Artiges2024]. 
 However, one critical aspect of the imaging pipeline which has seen limited open-source tool development is RF and gradient pulse design.  While RF pulse and gradient designers increasingly share code online
 in independent repositories, there are few sets of common pulse design tools maintained in a rigorous and consistent manner with easy-to-read code and tutorials. This is despite the reality that in many cases, carefully designed or application-specific RF and gradient pulses are crucial to the success of MRI or NMR techniques. An open
 source pulse design code library would facilitate the development and dissemination of
