### 2024-10-22 b109f68
Update paper.md to remove double parenthesis
