### 2024-10-22 c9f17c4
Update paper.bib correcting distorted reference
