+++ docs/paper/paper.bib
@@ -431,7 +431,7 @@ journal = {Magnetic Resonance in Medicine},
 keywords = {Bloch,RF,RF pulse design,Siegert shift,encoded MRI,field MRI,low,multiphoton,selective RF excitation},
 month = {apr},
 publisher = {John Wiley & Sons, Ltd},
-title = {{Selective excitation localized by the Bloch–Siegert shift and a B1+$$ {B}_1^{+} $$ gradient}},
+title = {{Selective excitation localized by the Bloch–Siegert shift and a B1+ gradient}},
 url = {https://onlinelibrary.wiley.com/doi/full/10.1002/mrm.29271 https://onlinelibrary.wiley.com/doi/abs/10.1002/mrm.29271 https://onlinelibrary.wiley.com/doi/10.1002/mrm.29271},
 year = {2022}
 }
