### 2024-10-22 730b135
Update paper.md with corrected figure reference
