### 2024-10-22 3d3a03f
Update paper.md with table caption
