### 2024-10-22 e051a48
Update paper.md pandoc style caption
