### 2024-10-22 4edbcca
Update paper.md test other convention for table
