### 2024-10-22 ec6371d
Update paper.md table captioning
