+++ docs/paper/paper.bib
@@ -288,6 +288,7 @@ author = {Duval, T and Leppert, Ilana R. and Cabana, Jean-Francois and Boudreau,
 booktitle = {Proc. Intl. Soc. Mag. Reson. Med.},
 pages = {2288},
 title = {{Quantitative MRI made easy with qMRLab}},
+doi = {10.13140/RG.2.2.25014.34881}, 
 year = {2018}
 }
 @article{SravanRavi2019,
