### 2024-10-22 #6586 @jonbmartin
@emdupre I have done all of the above.
- Changed title on Zenodo to specified
-  Expanded all acronyms in affiliations
- fixed broken in-text citations
- fixed malformed references
- added table caption

I have checked all of the SKIP DOIs and they are almost all conference proceedings from the ISMRM, which doesn't provide DOI. The one exception is the Michigan Image Reconstruction Toolbox reference, which has not been published. I double checked all of these references for publication on research-sharing platforms and could only find one for the qMRLab reference. I went ahead and added that, but it seems the others truly do not have DOIs.
