### 2024-10-22 b005ecb
Update paper.md make fig smaller
