### 2024-10-24 1e2c6e1
Update reference formatting in paper
