### 2024-10-24 #6586 @emdupre
Thank you, @jonbmartin  ! The paper now looks correct to me.

I did, though, just catch an issue on the archive : It seems that rather than editing the Zenodo archive, a new DOI was created in modifying the title. This is problematic as the version on the archive (v2) no longer matches the version on the software (v1.8.8).

Are you able to directly edit the existing archive, or create a new matching software version and archive version ?
