### 2024-10-24 #6586 @jonbmartin
Ah, I apologize for that! I just edited the existing archive (DOI: 10.5281/zenodo.13963346, URL: https://zenodo.org/records/13963346) to have the correct title. This is v1.8.8. Does that resolve the issue?
