### 2024-10-26 #6586 @emdupre
Thank you, @jonbmartin, for these edits ! The paper and archive now look correct to me.

I am now happy to recommend _PulPy_ to the EiC team for publication in _JOSS_ 🚀

Thank you, too, to @bwheelz36 and @curtcorum for your work in reviewing this submission ! _JOSS_ only works because of your work 💐
