### 2024-10-26 #6586 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
