### 2024-10-31 #6586 @Kevin-Mattheus-Moerman
@jonbmartin as AEiC for JOSS I will now help to process this submission for acceptance in JOSS. I will now process some final checks:

- [x] Software license tag listed here matches a tagged release

- [ ] Please edit the archive listing so that the author set matches the paper. Currently I think they may be down as contributors, but it would be good to list them all as authors.
